pyproject.toml
[tool.poetry]
name = "nanogpt"
version = "0.1.0"
description = "Experimentation with nanoGPT"
authors = ["Bartosz Nawrotek"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# torch = [
#     { version = "~2.1.0", source = "pypi", platform = "darwin" },
#     { version = "~2.1.0", source = "pytorch", platform = "linux" },
#     { version = "~2.1.0", source = "pytorch", platform = "win32" },
# ]
torch = { version = "~2.1.0", source = "pytorch" }
wandb = "^0.16.1"
portalocker = "^2.8.2"
lightning = "^2.1.3"
hydra-core = "^1.3.2"
pydantic = "^2.5.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pylint = "^3.0.3"
isort = "^5.13.2"
black = "^23.12.0"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu/"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
